+++ src/decorators/constants.ts
@@ -1,0 +1,3 @@
+export const EXPORTABLE_META = Symbol("nestbolt:exportable");
+export const EXPORT_COLUMNS_META = Symbol("nestbolt:export-columns");
+export const EXPORT_IGNORE_META = Symbol("nestbolt:export-ignore");