+++ src/decorators/exportable.decorator.ts
@@ -1,0 +1,9 @@
+import "reflect-metadata";
+import { EXPORTABLE_META } from "./constants";
+import type { ExportableOptions } from "./interfaces";
+
+export function Exportable(opts?: ExportableOptions): ClassDecorator {
+  return (target: Function) => {
+    Reflect.defineMetadata(EXPORTABLE_META, opts ?? {}, target);
+  };
+}